requirements.txt
amqp==2.6.1
asgiref==3.3.1
Babel==2.9.1
billiard==3.6.4.0
boto3==1.17.2
botocore==1.20.2
braintree==3.59.0
cairocffi==1.2.0
CairoSVG==2.5.2
celery==4.4.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.0.14
django-crispy-forms==1.11.2
django-getenv==1.3.2
django-phonenumber-field==5.2.0
django-storages==1.11.1
django-webp==2.0.0
django-webp-converter==0.3.0
djangorestframework==3.11.0
flower==0.9.3
gunicorn==20.1.0
html5lib==1.1
httpagentparser==1.9.1
idna==2.10
jmespath==0.10.0
kombu==4.6.11
phonenumbers==8.12.24
Pillow==8.1.0
pycparser==2.20
PyJWT==1.7.1
Pyphen==0.10.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
tornado==5.1.1
twilio==6.59.1
urllib3==1.26.3
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
